Renaming Python settings to capital P seems to make them work on Mac
For a long time I have had problems with the TAB setting in Python on Windows,
not getting spaces, except if there already were spaces for indentation.
Using the Windows settings (which had diverged) made that fail on Mac as well.
Capitalizing the settings file name made it work on Mac, not sure about windows
yet.